New formula: `git-wip`
git-wip is a script that will manage Work In Progress (or WIP) branches. WIP
branches are mostly throw away but identify points of development between
commits. The intent is to tie this script into your editor so that each time
you save your file, the git-wip script captures that state in git. git-wip
also helps you return back to a previous state of development.